// web/src/lib/api/generated/models/githubComMantonxViewraInternalApplicationAuthUserResponse.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * ViewRA Media Server API
 * Self-hosted media server for movies, TV shows, and music
 * OpenAPI spec version: 0.0.1
 */

export interface GithubComMantonxViewraInternalApplicationAuthUserResponse {
  created_at?: string
  display_name?: string
  id?: string
  is_admin?: boolean
  username?: string
}